Refactor files node API for more code reuse
