+++ tests/lib/Files/Node/NodeTest.php
@@ -9,14 +9,23 @@
 namespace Test\Files\Node;
 
 use OC\Files\FileInfo;
+use OCP\Files\NotFoundException;
 
-class NodeTest extends \Test\TestCase {
-	private $user;
+/**
+ * Class NodeTest
+ *
+ * @package Test\Files\Node
+ */
+abstract class NodeTest extends \Test\TestCase {
+	protected $viewDeleteMethod = 'unlink';
+	protected $user;
 
 	protected function setUp() {
 		parent::setUp();
 		$this->user = new \OC\User\User('', new \Test\Util\User\Dummy);
 	}
+
+	protected abstract function createTestNode($root, $view, $path);
 
 	protected function getMockStorage() {
 		$storage = $this->createMock('\OCP\Files\Storage');
@@ -29,6 +38,119 @@ class NodeTest extends \Test\TestCase {
 	protected function getFileInfo($data) {
 		return new FileInfo('', $this->getMockStorage(), '', $data, null);
 	}
+
+	public function testDelete() {
+		$manager = $this->createMock('\OC\Files\Mount\Manager');
+
+		/**
+		 * @var \OC\Files\View | \PHPUnit_Framework_MockObject_MockObject $view
+		 */
+		$view = $this->createMock('\OC\Files\View');
+
+		$root = $this->createMock('\OC\Files\Node\Root', [], [$manager, $view, $this->user]);
+		$root->expects($this->exactly(2))
+			->method('emit')
+			->will($this->returnValue(true));
+		$root->expects($this->any())
+			->method('getUser')
+			->will($this->returnValue($this->user));
+
+		$view->expects($this->once())
+			->method('getFileInfo')
+			->with('/bar/foo')
+			->will($this->returnValue($this->getFileInfo(['permissions' => \OCP\Constants::PERMISSION_ALL])));
+
+		$view->expects($this->once())
+			->method($this->viewDeleteMethod)
+			->with('/bar/foo')
+			->will($this->returnValue(true));
+
+		$node = $this->createTestNode($root, $view, '/bar/foo');
+		$node->delete();
+	}
+
+	public function testDeleteHooks() {
+		$test = $this;
+		$hooksRun = 0;
+		/**
+		 * @param \OC\Files\Node\File $node
+		 */
+		$preListener = function ($node) use (&$test, &$hooksRun) {
+			$test->assertInstanceOf($this->nodeClass, $node);
+			$test->assertEquals('foo', $node->getInternalPath());
+			$test->assertEquals('/bar/foo', $node->getPath());
+			$test->assertEquals(1, $node->getId());
+			$hooksRun++;
+		};
+
+		/**
+		 * @param \OC\Files\Node\File $node
+		 */
+		$postListener = function ($node) use (&$test, &$hooksRun) {
+			$test->assertInstanceOf($this->nonExistingNodeClass, $node);
+			$test->assertEquals('foo', $node->getInternalPath());
+			$test->assertEquals('/bar/foo', $node->getPath());
+			$test->assertEquals(1, $node->getId());
+			$test->assertEquals('text/plain', $node->getMimeType());
+			$hooksRun++;
+		};
+
+		/**
+		 * @var \OC\Files\Mount\Manager $manager
+		 */
+		$manager = $this->createMock('\OC\Files\Mount\Manager');
+		/**
+		 * @var \OC\Files\View | \PHPUnit_Framework_MockObject_MockObject $view
+		 */
+		$view = $this->createMock('\OC\Files\View');
+		$root = new \OC\Files\Node\Root($manager, $view, $this->user);
+		$root->listen('\OC\Files', 'preDelete', $preListener);
+		$root->listen('\OC\Files', 'postDelete', $postListener);
+
+		$view->expects($this->any())
+			->method('getFileInfo')
+			->with('/bar/foo')
+			->will($this->returnValue($this->getFileInfo(['permissions' => \OCP\Constants::PERMISSION_ALL, 'fileid' => 1, 'mimetype' => 'text/plain'])));
+
+		$view->expects($this->once())
+			->method($this->viewDeleteMethod)
+			->with('/bar/foo')
+			->will($this->returnValue(true));
+
+		$view->expects($this->any())
+			->method('resolvePath')
+			->with('/bar/foo')
+			->will($this->returnValue([null, 'foo']));
+
+		$node = $this->createTestNode($root, $view, '/bar/foo');
+		$node->delete();
+		$this->assertEquals(2, $hooksRun);
+	}
+
+	/**
+	 * @expectedException \OCP\Files\NotPermittedException
+	 */
+	public function testDeleteNotPermitted() {
+		$manager = $this->createMock('\OC\Files\Mount\Manager');
+		/**
+		 * @var \OC\Files\View | \PHPUnit_Framework_MockObject_MockObject $view
+		 */
+		$view = $this->createMock('\OC\Files\View');
+		$root = $this->createMock('\OC\Files\Node\Root', [], [$manager, $view, $this->user]);
+
+		$root->expects($this->any())
+			->method('getUser')
+			->will($this->returnValue($this->user));
+
+		$view->expects($this->once())
+			->method('getFileInfo')
+			->with('/bar/foo')
+			->will($this->returnValue($this->getFileInfo(['permissions' => \OCP\Constants::PERMISSION_READ])));
+
+		$node = $this->createTestNode($root, $view, '/bar/foo');
+		$node->delete();
+	}
+
 
 	public function testStat() {
 		$manager = $this->createMock('\OC\Files\Mount\Manager');
@@ -54,7 +176,7 @@ class NodeTest extends \Test\TestCase {
 			->with('/bar/foo')
 			->will($this->returnValue($stat));
 
-		$node = new \OC\Files\Node\File($root, $view, '/bar/foo');
+		$node = $this->createTestNode($root, $view, '/bar/foo');
 		$this->assertEquals($stat, $node->stat());
 	}
 
@@ -81,7 +203,7 @@ class NodeTest extends \Test\TestCase {
 			->with('/bar/foo')
 			->will($this->returnValue($stat));
 
-		$node = new \OC\Files\Node\File($root, $view, '/bar/foo');
+		$node = $this->createTestNode($root, $view, '/bar/foo');
 		$this->assertEquals(1, $node->getId());
 	}
 
@@ -109,7 +231,7 @@ class NodeTest extends \Test\TestCase {
 			->with('/bar/foo')
 			->will($this->returnValue($stat));
 
-		$node = new \OC\Files\Node\File($root, $view, '/bar/foo');
+		$node = $this->createTestNode($root, $view, '/bar/foo');
 		$this->assertEquals(100, $node->getSize());
 	}
 
@@ -136,7 +258,7 @@ class NodeTest extends \Test\TestCase {
 			->with('/bar/foo')
 			->will($this->returnValue($stat));
 
-		$node = new \OC\Files\Node\File($root, $view, '/bar/foo');
+		$node = $this->createTestNode($root, $view, '/bar/foo');
 		$this->assertEquals('qwerty', $node->getEtag());
 	}
 
@@ -163,7 +285,7 @@ class NodeTest extends \Test\TestCase {
 			->with('/bar/foo')
 			->will($this->returnValue($stat));
 
-		$node = new \OC\Files\Node\File($root, $view, '/bar/foo');
+		$node = $this->createTestNode($root, $view, '/bar/foo');
 		$this->assertEquals(50, $node->getMTime());
 	}
 
@@ -188,7 +310,7 @@ class NodeTest extends \Test\TestCase {
 			->will($this->returnValue([$storage, 'foo']));
 
 
-		$node = new \OC\Files\Node\File($root, $view, '/bar/foo');
+		$node = $this->createTestNode($root, $view, '/bar/foo');
 		$this->assertEquals($storage, $node->getStorage());
 	}
 
@@ -203,7 +325,7 @@ class NodeTest extends \Test\TestCase {
 			->method('getUser')
 			->will($this->returnValue($this->user));
 
-		$node = new \OC\Files\Node\File($root, $view, '/bar/foo');
+		$node = $this->createTestNode($root, $view, '/bar/foo');
 		$this->assertEquals('/bar/foo', $node->getPath());
 	}
 
@@ -228,7 +350,7 @@ class NodeTest extends \Test\TestCase {
 			->will($this->returnValue([$storage, 'foo']));
 
 
-		$node = new \OC\Files\Node\File($root, $view, '/bar/foo');
+		$node = $this->createTestNode($root, $view, '/bar/foo');
 		$this->assertEquals('foo', $node->getInternalPath());
 	}
 
@@ -243,7 +365,7 @@ class NodeTest extends \Test\TestCase {
 			->method('getUser')
 			->will($this->returnValue($this->user));
 
-		$node = new \OC\Files\Node\File($root, $view, '/bar/foo');
+		$node = $this->createTestNode($root, $view, '/bar/foo');
 		$this->assertEquals('foo', $node->getName());
 	}
 
@@ -268,7 +390,7 @@ class NodeTest extends \Test\TestCase {
 			->with('/bar/foo')
 			->will($this->returnValue($this->getFileInfo(['permissions' => \OCP\Constants::PERMISSION_ALL])));
 
-		$node = new \OC\Files\Node\Node($root, $view, '/bar/foo');
+		$node = $this->createTestNode($root, $view, '/bar/foo');
 		$node->touch(100);
 		$this->assertEquals(100, $node->getMTime());
 	}
@@ -321,7 +443,7 @@ class NodeTest extends \Test\TestCase {
 			->with('/bar/foo')
 			->will($this->returnValue($this->getFileInfo(['permissions' => \OCP\Constants::PERMISSION_ALL])));
 
-		$node = new \OC\Files\Node\Node($root, $view, '/bar/foo');
+		$node = $this->createTestNode($root, $view, '/bar/foo');
 		$node->touch(100);
 		$this->assertEquals(2, $hooksRun);
 	}
@@ -345,7 +467,7 @@ class NodeTest extends \Test\TestCase {
 			->with('/bar/foo')
 			->will($this->returnValue($this->getFileInfo(['permissions' => \OCP\Constants::PERMISSION_READ])));
 
-		$node = new \OC\Files\Node\Node($root, $view, '/bar/foo');
+		$node = $this->createTestNode($root, $view, '/bar/foo');
 		$node->touch(100);
 	}
 
@@ -360,7 +482,264 @@ class NodeTest extends \Test\TestCase {
 		$view = $this->createMock('\OC\Files\View');
 		$root = $this->createMock('\OC\Files\Node\Root', [], [$manager, $view, $this->user]);
 
-		$node = new \OC\Files\Node\Node($root, $view, '/../foo');
+		$node = $this->createTestNode($root, $view, '/../foo');
 		$node->getFileInfo();
+	}
+
+	public function testCopySameStorage() {
+		/**
+		 * @var \OC\Files\Mount\Manager $manager
+		 */
+		$manager = $this->createMock('\OC\Files\Mount\Manager');
+		/**
+		 * @var \OC\Files\View | \PHPUnit_Framework_MockObject_MockObject $view
+		 */
+		$view = $this->createMock('\OC\Files\View');
+		$root = $this->createMock('\OC\Files\Node\Root', [], [$manager, $view, $this->user]);
+
+		$view->expects($this->any())
+			->method('copy')
+			->with('/bar/foo', '/bar/asd')
+			->will($this->returnValue(true));
+
+		$view->expects($this->any())
+			->method('getFileInfo')
+			->will($this->returnValue($this->getFileInfo(['permissions' => \OCP\Constants::PERMISSION_ALL, 'fileid' => 3])));
+
+		$node = $this->createTestNode($root, $view, '/bar/foo');
+		$parentNode = new \OC\Files\Node\Folder($root, $view, '/bar');
+		$newNode = $this->createTestNode($root, $view, '/bar/asd');
+
+		$root->expects($this->exactly(2))
+			->method('get')
+			->will($this->returnValueMap([
+				['/bar/asd', $newNode],
+				['/bar', $parentNode]
+			]));
+
+		$target = $node->copy('/bar/asd');
+		$this->assertInstanceOf($this->nodeClass, $target);
+		$this->assertEquals(3, $target->getId());
+	}
+
+	/**
+	 * @expectedException \OCP\Files\NotPermittedException
+	 */
+	public function testCopyNotPermitted() {
+		/**
+		 * @var \OC\Files\Mount\Manager $manager
+		 */
+		$manager = $this->createMock('\OC\Files\Mount\Manager');
+		/**
+		 * @var \OC\Files\View | \PHPUnit_Framework_MockObject_MockObject $view
+		 */
+		$view = $this->createMock('\OC\Files\View');
+		$root = $this->createMock('\OC\Files\Node\Root', [], [$manager, $view, $this->user]);
+		/**
+		 * @var \OC\Files\Storage\Storage | \PHPUnit_Framework_MockObject_MockObject $storage
+		 */
+		$storage = $this->createMock('\OC\Files\Storage\Storage');
+
+		$root->expects($this->never())
+			->method('getMount');
+
+		$storage->expects($this->never())
+			->method('copy');
+
+		$view->expects($this->any())
+			->method('getFileInfo')
+			->will($this->returnValue($this->getFileInfo(['permissions' => \OCP\Constants::PERMISSION_READ, 'fileid' => 3])));
+
+		$node = $this->createTestNode($root, $view, '/bar/foo');
+		$parentNode = new \OC\Files\Node\Folder($root, $view, '/bar');
+
+		$root->expects($this->once())
+			->method('get')
+			->will($this->returnValueMap([
+				['/bar', $parentNode]
+			]));
+
+		$node->copy('/bar/asd');
+	}
+
+	/**
+	 * @expectedException \OCP\Files\NotFoundException
+	 */
+	public function testCopyNoParent() {
+		/**
+		 * @var \OC\Files\Mount\Manager $manager
+		 */
+		$manager = $this->createMock('\OC\Files\Mount\Manager');
+		/**
+		 * @var \OC\Files\View | \PHPUnit_Framework_MockObject_MockObject $view
+		 */
+		$view = $this->createMock('\OC\Files\View');
+		$root = $this->createMock('\OC\Files\Node\Root', [], [$manager, $view, $this->user]);
+
+		$view->expects($this->never())
+			->method('copy');
+
+		$node = $this->createTestNode($root, $view, '/bar/foo');
+
+		$root->expects($this->once())
+			->method('get')
+			->with('/bar/asd')
+			->will($this->throwException(new NotFoundException()));
+
+		$node->copy('/bar/asd/foo');
+	}
+
+	/**
+	 * @expectedException \OCP\Files\NotPermittedException
+	 */
+	public function testCopyParentIsFile() {
+		/**
+		 * @var \OC\Files\Mount\Manager $manager
+		 */
+		$manager = $this->createMock('\OC\Files\Mount\Manager');
+		/**
+		 * @var \OC\Files\View | \PHPUnit_Framework_MockObject_MockObject $view
+		 */
+		$view = $this->createMock('\OC\Files\View');
+		$root = $this->createMock('\OC\Files\Node\Root', [], [$manager, $view, $this->user]);
+
+		$view->expects($this->never())
+			->method('copy');
+
+		$node = $this->createTestNode($root, $view, '/bar/foo');
+		$parentNode = new \OC\Files\Node\File($root, $view, '/bar');
+
+		$root->expects($this->once())
+			->method('get')
+			->will($this->returnValueMap([
+				['/bar', $parentNode]
+			]));
+
+		$node->copy('/bar/asd');
+	}
+
+	public function testMoveSameStorage() {
+		/**
+		 * @var \OC\Files\Mount\Manager $manager
+		 */
+		$manager = $this->createMock('\OC\Files\Mount\Manager');
+		/**
+		 * @var \OC\Files\View | \PHPUnit_Framework_MockObject_MockObject $view
+		 */
+		$view = $this->createMock('\OC\Files\View');
+		$root = $this->createMock('\OC\Files\Node\Root', [], [$manager, $view, $this->user]);
+
+		$view->expects($this->any())
+			->method('rename')
+			->with('/bar/foo', '/bar/asd')
+			->will($this->returnValue(true));
+
+		$view->expects($this->any())
+			->method('getFileInfo')
+			->will($this->returnValue($this->getFileInfo(['permissions' => \OCP\Constants::PERMISSION_ALL, 'fileid' => 1])));
+
+		$node = $this->createTestNode($root, $view, '/bar/foo');
+		$parentNode = new \OC\Files\Node\Folder($root, $view, '/bar');
+
+		$root->expects($this->any())
+			->method('get')
+			->will($this->returnValueMap([['/bar', $parentNode], ['/bar/asd', $node]]));
+
+		$target = $node->move('/bar/asd');
+		$this->assertInstanceOf($this->nodeClass, $target);
+		$this->assertEquals(1, $target->getId());
+		$this->assertEquals('/bar/asd', $node->getPath());
+	}
+
+	/**
+	 * @expectedException \OCP\Files\NotPermittedException
+	 */
+	public function testMoveNotPermitted() {
+		/**
+		 * @var \OC\Files\Mount\Manager $manager
+		 */
+		$manager = $this->createMock('\OC\Files\Mount\Manager');
+		/**
+		 * @var \OC\Files\View | \PHPUnit_Framework_MockObject_MockObject $view
+		 */
+		$view = $this->createMock('\OC\Files\View');
+		$root = $this->createMock('\OC\Files\Node\Root', [], [$manager, $view, $this->user]);
+
+		$view->expects($this->any())
+			->method('getFileInfo')
+			->will($this->returnValue($this->getFileInfo(['permissions' => \OCP\Constants::PERMISSION_READ])));
+
+		$view->expects($this->never())
+			->method('rename');
+
+		$node = $this->createTestNode($root, $view, '/bar/foo');
+		$parentNode = new \OC\Files\Node\Folder($root, $view, '/bar');
+
+		$root->expects($this->once())
+			->method('get')
+			->with('/bar')
+			->will($this->returnValue($parentNode));
+
+		$node->move('/bar/asd');
+	}
+
+	/**
+	 * @expectedException \OCP\Files\NotFoundException
+	 */
+	public function testMoveNoParent() {
+		/**
+		 * @var \OC\Files\Mount\Manager $manager
+		 */
+		$manager = $this->createMock('\OC\Files\Mount\Manager');
+		/**
+		 * @var \OC\Files\View | \PHPUnit_Framework_MockObject_MockObject $view
+		 */
+		$view = $this->createMock('\OC\Files\View');
+		$root = $this->createMock('\OC\Files\Node\Root', [], [$manager, $view, $this->user]);
+		/**
+		 * @var \OC\Files\Storage\Storage | \PHPUnit_Framework_MockObject_MockObject $storage
+		 */
+		$storage = $this->createMock('\OC\Files\Storage\Storage');
+
+		$storage->expects($this->never())
+			->method('rename');
+
+		$node = $this->createTestNode($root, $view, '/bar/foo');
+		$parentNode = new \OC\Files\Node\Folder($root, $view, '/bar');
+
+		$root->expects($this->once())
+			->method('get')
+			->with('/bar')
+			->will($this->throwException(new NotFoundException()));
+
+		$node->move('/bar/asd');
+	}
+
+	/**
+	 * @expectedException \OCP\Files\NotPermittedException
+	 */
+	public function testMoveParentIsFile() {
+		/**
+		 * @var \OC\Files\Mount\Manager $manager
+		 */
+		$manager = $this->createMock('\OC\Files\Mount\Manager');
+		/**
+		 * @var \OC\Files\View | \PHPUnit_Framework_MockObject_MockObject $view
+		 */
+		$view = $this->createMock('\OC\Files\View');
+		$root = $this->createMock('\OC\Files\Node\Root', [], [$manager, $view, $this->user]);
+
+		$view->expects($this->never())
+			->method('rename');
+
+		$node = $this->createTestNode($root, $view, '/bar/foo');
+		$parentNode = new \OC\Files\Node\File($root, $view, '/bar');
+
+		$root->expects($this->once())
+			->method('get')
+			->with('/bar')
+			->will($this->returnValue($parentNode));
+
+		$node->move('/bar/asd');
 	}
 }
